Implemented Write*BE functions

+++ stratigraphy/nbt/IOUtils.h
@@ -8,11 +8,11 @@
 namespace stratigraphy { namespace nbt {
     void WriteShortBE(std::ostream& o, short s, char *buff);
     void WriteIntBE  (std::ostream& o, int   i, char *buff);
-    void WriteLongBE (std::ostream& o, long  l, char *buff);
+    void WriteLongBE (std::ostream& o, long long l, char *buff);
     
     short ReadShortBE(std::istream& from, char *buff);
     int   ReadIntBE  (std::istream& from, char *buff);
-    long  ReadLongBE (std::istream& from, char *buff);
+    long long ReadLongBE (std::istream& from, char *buff);
 }; };
 
 #endif
